% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_control.R
\name{aucs}
\alias{aucs}
\title{Sets the performance metric to be area under the AUC and PR curve}
\usage{
aucs(data, lev = NULL, model = NULL)
}
\value{
named vector of two functions: \code{prAUC} or \code{rocAUC}
}
\description{
Sets the performance metric to be area under the AUC and PR curve
}
\details{
Modifies \code{summaryFunction} used in \code{\link[caret]{trainControl}}. 
Creates additional summary functions for \code{summaryFunction}. 

Possible methods for evaluating cross-validation performance in \code{\link[caret]{trainControl}} 
can then be \code{prAUC} or \code{rocAUC} for area under the precision-recall curve 
and area under the ROC curve, respectively. 

Their performance names are found in \code{\link{getPerformanceNames}}.
}
\examples{
# not run
trainControl(method='cv', classProbs=T, index = index, summaryFunction=aucs)
}
