% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{cleanData}
\alias{cleanData}
\title{Clean data to desired format ready for 'arulesSequences'.}
\usage{
cleanData(event, tType = "r", removeChemo = T)
}
\arguments{
\item{event:}{dataframe, events as rows and at least the following columns:
\itemize{
\item \code{iois} patient id, int
\item \code{agent} event type, char
\item \code{agent_detail} event attribute, char
\item \code{startdate} event start date, char will be formatted 'YYYY-MM-DD' Date
\item \code{enddate} event end date, char will be formatted 'YYYY-MM-DD' Date
}}

\item{tType:}{char, type of tumor data to use: see 'getVolTypeName' for valid values}

\item{removeChemo:}{bool, remove chemo event types}
}
\value{
event: dataframe, cleaned and formated information with added columns:
\itemize{
\item \code{eventID} event id, num, see \code{\link{resetEventIDsToDays}}
\item \code{endEventID} end event id, num, see \code{\link{resetEventIDsToDays}}
\item \code{eventName} unique event name, char, see \code{\link{createEventNames}}
}
}
\description{
Clean data to desired format ready for 'arulesSequences'.
}
\details{

}
\examples{
data("fake_data")
cleaned <- cleanData(fake_data$events, tType='r',  removeChemo=T)

}
