% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{createEventNames}
\alias{createEventNames}
\title{Create unique event names for each attribute of an event type.}
\usage{
createEventNames(event, tType, removeChemo)
}
\arguments{
\item{tType}{char, shortcut for type of tumor volume information}

\item{event:}{dataframe, all events across all patients, each row is an event, returned from 'getData' in 'getData.R'}

\item{removeChemo:}{bool, remove chemo event types}
}
\value{
event: dataframe, cleaned and formated information that is ammendable to SPM
}
\description{
Event names serve as an unique event in SPM. This will merge agent and agent_detail into a new column, eventName
Here, each attribute of an event type is stated in the event name. Events with continuous values will first discretized.
(e.g., "tumorINCR", tumor volume is the event from agent column, and INCR is the attribute converted from agent_detail column.)
}
\details{
\code{event} contains events as rows and at least the following columns:
\itemize{
\item \code{iois} patient id, char
\item \code{agent} event type, char
\item \code{agent_detail} event attribute, char
\item \code{startdate} event start date in the format 'YYYY-MM-DD', char
\item \code{enddate} event end date in the format 'YYYY-MM-DD', char
\item \code{eventID} event id
\item \code{eventName} distinct names for an event (e.g., mentalStatus0: agent is mental status, agent_detail is 0)
}
}
\examples{
data("fake_data")
events <- fake_data$events
events$startdate <- as.Date(as.character(events$startdate), format='\%Y-\%m-\%d')
events$enddate <- as.Date(as.character(events$enddate), format='\%Y-\%m-\%d')

cleaned <- removeData(events)
cleaned <- resetEventIDsToDays(cleaned)
cleaned <- removePreRadiation(cleaned)

cleaned <- createEventNames(event=cleaned,
                            tType='r',
                            removeChemo=T)
                            
}
