% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_getPatterns.R
\name{createTransactions}
\alias{createTransactions}
\title{Convert events into transactions.}
\usage{
createTransactions(event, filename)
}
\arguments{
\item{event}{dataframe of events,
where rows are single events, and
least the following columns:
\itemize{
\item \code{iois} patient id
\item \code{eventID} event id, see \code{\link{resetEventIDsToDays}}
\item \code{eventName} unique event name, see \code{\link{createEventNames}}
}}

\item{filename}{path to save transactions file}
}
\description{
Convert events into transactions.
}
\details{
Creates a transaction file with the following 'n' columns: \itemize{
\item column 1: sequenceID
\item column 2: timeStamp/eventID
\item column 3 = number of items
\item columns 4-n = items in event
}

Transaction file is to be read by \code{\link[arulesSequences]{read_baskets}}.
}
\examples{
data("fake_data")
outputDir <- '~/test'

# save tumor location and laterility strings before event cleaning
fake_tumorInfo <- fake_data$events 
fake_demo <- fake_data$demo

# clean data
fake_data$events <- cleanData(fake_data$events, tType = 'rate')
cat('...',nrow(fake_data$events), " events left for SPM after cleaning", '\\n')

# collect patient info for each event
fake_data <- merge(fake_data$events, fake_data$person, by='iois', all.x=T) 

# prep for each event, since age does change
# get survival labels, these also change
fake_data <- prepDemographics(fake_data, fake_demo) 
fake_data <- prepSurvivalLabels(fake_data) 

# get first tumor location
fake_data <- getTumorLocation(fake_data, fake_tumorInfo) 

# transactions
trans <- createTransactions(event = fake_data, 
                            filename = file.path(outputDir, 'example_transactions.txt'))
}
