% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_discretize.R
\name{discretizeKPS}
\alias{discretizeKPS}
\title{Discretize KPS according to expert opinion}
\usage{
discretizeKPS(kps)
}
\arguments{
\item{kps}{a dataframe of KPS scores, with columns: \itemize{
\item \code{startdate} date of KPS evaluation)
\item \code{agent_detail} KPS}}
}
\value{
A dataframe with original data \code{kps} 
plus discretized KPS categories in \code{eventName} column.
}
\description{
KPS is categorized into UNCH, SigDECR, SigINCR, DECR, and INCR.
}
\details{
KPS categories are \itemize{
\item initial: first KPS in sequence
\item UNCH: unchanged, KPS is the same as the previous KPS
\item SigDECR: significant decrease, KPS drops to <= 60 OR KPS was already <= 60 and drops further
\item SigINCR: significant increase, KPS is > 60 AND previous KPS was <=60
\item DECR: decrease, KPS drops but remains above 60 AND psrevious KPS was above 60
\item INCR: increase, KPS improves AND is above 60 AND previous KPS was above 60
}
}
\examples{
data("fake_data")

# get one person's set of KPS exams
kps_set <- fake_data$events[fake_data$events$agent=='KPS' & fake_data$events$iois=='2' , ]

# convert to categorical
kps_set_discretized <- discretizeKPS(kps_set)
}
