% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_discretize.R
\name{discretizeTumVols}
\alias{discretizeTumVols}
\title{Discretize tumor volume.}
\usage{
discretizeTumVols(tumorVols, tType)
}
\arguments{
\item{tumorVols}{a dataframe of tumor volume, see Details}

\item{tType}{char, abbrev. type of tumor data to use, see \code{\link{getVolTypeName}} for valid values}
}
\value{
A dataframe with original data in \code{tumorVols} plus discretized volumes in \code{eventName} column.
}
\description{
Discretize tumor volume.
}
\details{
\code{tumorVols} contains tumor measurement events as rows and at least the following columns: \itemize{
\item iois = patient id, char
\item agent_detail = event attribute, i.e., measured tumor volume in mm3 char
\item startdate = event start datein the format 'YYYY-MM-DD', char
\item eventName = event name, char
}
}
\examples{
data("fake_data")
vols <- fake_data$events[fake_data$events$agent=='measurement', ]

# format data
vols$startdate <- as.Date(as.character(vols$startdate), format='\%Y-\%m-\%d')
vols$enddate <- as.Date(as.character(vols$enddate), format='\%Y-\%m-\%d')
vols <- resetEventIDsToDays(vols)
vols <- removeData(vols) # get rid of excess details and empty or no data

vols_dist <- discretizeTumVols(vols, tType = 'rate')
}
