% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_myHelpers.R
\name{findPatternDays}
\alias{findPatternDays}
\title{Find People with a sequence pattern in their events}
\usage{
findPatternDays(patterns, events, maxgap)
}
\value{
returns a list of patterns, where each pattern has a list of patients, and each patient has a list of days when event occurred (eventID)
}
\description{
Find People with a sequence pattern in their events
}
\examples{
data("fake_data")
fake_tumorInfo <- fake_data$events # save tumor location and laterility strings before event cleaning
fake_demo <- fake_data$demo
fake_data$events <- cleanData(fake_data$events, tType = 'rate')
cat('...',nrow(fake_data$events), " events left for SPM after cleaning", '\\n')

# collect patient info for each event
# note that only eventName, iois, and eventID columns are used for SPM
fake_data <- merge(fake_data$events, fake_data$person, by='iois', all.x=T)
fake_data <- prepDemographics(fake_data, fake_demo) # prep for each event, since age does change
fake_data <- prepSurvivalLabels(fake_data) # get survival labels, these also change
fake_data <- getTumorLocation(fake_data, fake_tumorInfo) # get first tumor location

# pattern
pSPM <- getSeqPatterns(data = fake_data,
                       transFilename = 'example_transactions.txt',
                       createT = T,
                       supp = 0.2,
                       maxgap = 60,
                       maxlen = 2,
                       maxsize = 2)
pSPM$patterns <- as(pSPM$freqseq, "data.frame")
pSPM$patterns$sequence <- as.character(pSPM$patterns$sequence)

# days when pattern occur
patternDays <- findPatternDays(pSPM$patterns, pSPM$data, maxgap=60) 

}
