% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{getAge}
\alias{getAge}
\title{Calculate numeric age.}
\usage{
getAge(birthDate, date)
}
\arguments{
\item{birthDate:}{date of birth in format 'YYYY-MM-DD', date}

\item{date:}{current date to evaluate age in format 'YYYY-MM-DD', date}
}
\value{
\code{age}, numeric age at \code{date}
}
\description{
Calculate numeric age.
}
\examples{
data("fake_data")
birthdates <- fake_data$demo$DOB
today <- as.Date('2000-10-30')
ages <- getAge(birthdates, date=today)

}
