% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_discretize.R
\name{getAgeDecade}
\alias{getAgeDecade}
\title{Get age decade.}
\usage{
getAgeDecade(birthDate, date, categories)
}
\arguments{
\item{birthDate}{date of birth, in the format 'YYYY-MM-DD', chars will be converted to Date}

\item{date}{date to evaluate age decade, in the format 'YYYY-MM-DD', chars will be converted to Date}

\item{categories}{ number of categories or a vector with boundaries 
    (all values outside the boundaries will be set to NA). }
}
\value{
age decade
}
\description{
Get age decade.
}
\examples{

}
\seealso{
\code{\link[arules]{discretize}}
}
