% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{getChemoOverlapEvents}
\alias{getChemoOverlapEvents}
\title{Get events that overlap with chemotherapy, specifically just the top 5 most used drug.}
\usage{
getChemoOverlapEvents(event, tType)
}
\description{
args
- event: dataframe of all events called from Agile database (getData())
}
\details{
returns
- overlaps: dataframe, events that occurred during chemotherapy
}
