% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_control.R
\name{getCV}
\alias{getCV}
\title{Perform cross-validation on training partition.}
\usage{
getCV(data, formula, labelName, lasso = FALSE, llength = NULL,
  lmax = NULL, predictors = NULL, needToRemove = NULL,
  createModelMatrix = FALSE, metric = c("prAUC", "rocAUC"), seed, folds,
  verbose = TRUE)
}
\arguments{
\item{data}{dataframe, rows are samples, cols are features plus some metadata
not meant for modeling and will be removed}

\item{formula}{char or formula object}

\item{labelName}{char, column name of binary label}

\item{lasso}{logical, whether to use lasso regularization}

\item{llength}{num, number of lambdas to consider up to \code{lmax}}

\item{lmax}{num, maximum lambda to consider, cannot be NULL if lambda is NULL}

\item{predictors}{char, names of columns in \code{data} that should be in logit fit data}

\item{needToRemove}{char, names of columns in \code{data} that should not be in logit fit data}

\item{createModelMatrix}{logical, call \code{\link[stats]{model.matrix}}}

\item{metric}{char, see \code{aucs}}

\item{seed}{int, seed for split}

\item{folds}{number of folds}

\item{verbose}{logical, True for print, False for silence}
}
\value{
\code{folds} a list of arrays indicating row 
position integers corresponding to fold split
}
\description{
Perform cross-validation on training partition.
}
\details{
Removes rows for which a column is NA.
}
\seealso{
\code{\link[caret]{groupKFold}},
\code{\link{getPatIDs}},
\code{\link{getClassLabels}},
\code{\link{getData}}
}
