% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_callDatabase.R
\name{getData}
\alias{getData}
\alias{getEventData}
\alias{getSurvData}
\alias{getDemographics}
\title{Get patient event and survival data from a database}
\usage{
getData(database, eventTable, personTable)

getEventData(database, eventTable)

getSurvData(database, personTable)

getDemographics(database, personTable)
}
\arguments{
\item{database}{a MySQL database, where connection is
authenticated through a configuration file with group named \code{database})}

\item{eventTable}{an event-based table in \code{database}, containing at least:
\itemize{
\item iois = patient id, char
\item agent = event type, char
\item agent_detail = event attribute, char
\item startdate = event start datein the format 'YYYY-MM-DD', char
\item enddate = event end date in the format 'YYYY-MM-DD', char
\item recurrenceNum = patient's recurrence count (since baseline) during event, int
\item daysSinceBaseline = days since baseline, int
}}

\item{personTable}{a patient-based table in \code{database}, containing at least:
\itemize{
\item iois = patient id, char
\item survival = total survival days at time of data collection, int
\item status = patient status at the end of data collection , 'DECEASED' or 'ALIVE', char
\item DOB = date of birth in the format 'YYYY-MM-DD', char
\item gender = gender, char
\item ethnicity = ethnicity, char
\item MSP = MGMT promoter methytation status, char ('U' = unmethylated, 'M' = methylated)
\item IDH1 = status of IDH1 gene, char ('WT' = wild type, 'MUT' = mutated)
}}
}
\value{
\code{getData} returns a named list containing two dataframes, 'person' for patient data 
(returned by \code{getSurvData}) and 'event' for event data (returned by \code{getEventData})
}
\description{
Get patient event and survival data from a database

\code{getEventData}: get event data from \code{eventTable} in database

\code{getSurvData}: get patient survival days and survival status from \code{personTable} in database

\code{getDemographics}: get patient demographics and two biomarkers from \code{personTable} in database
}
\details{
Note: tumor lateriality and location is searched for in agent_detail regardless of agent/event type.
}
