% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{getLaterality}
\alias{getLaterality}
\title{Parse the tumor location string to get laterality.}
\usage{
getLaterality(str)
}
\arguments{
\item{str:}{tumor location string, char}
}
\value{
tumor laterality, char
}
\description{
Parse the tumor location string to get laterality.
}
\examples{
data("fake_data")

# do for all patients
lat_strs <- fake_data$events[fake_data$events$agent=='Surgery',] # event has tumor location information
lat_strs$laterality <- unlist(lapply(lat_strs$agent_detail, function(i) getLaterality(i) ))

# single patient
lat_str <- fake_data$events[fake_data$events$agent=='Surgery',][1,]
lat <- getLaterality(lat_str$agent_detail)

}
