% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{getLocation}
\alias{getLocation}
\title{Parse the tumor location string to get location}
\usage{
getLocation(str)
}
\arguments{
\item{str:}{tumor location string, char}
}
\value{
tumor location, char
}
\description{
Parse the tumor location string to get location
}
\details{
Possible locatios are  c('frontal','temporal', 'occipital','parietal','cerebellar', 'thalamic', 'sellar',
'corpus callosum', 'pineal', 'midbrain').
}
\examples{
data("fake_data")

# single patient
loc_str <- fake_data$events[fake_data$events$agent=='Surgery',][1,]
loc <- getLocation(loc_str$agent_detail)

}
