% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_control.R
\name{getModelStruct}
\alias{getModelStruct}
\title{Get logit coefficients}
\usage{
getModelStruct(fit)
}
\arguments{
\item{fit}{logit fit}
}
\description{
Assumes \code{fit} contains the following (returned from \code{\link[caret]{trainControl}}):
\itemize{
\item \code{traininData} data used for training, with \code{.outcome} as the classification label column name
\item \code{fintalMode} the retrained, final model using selected \code{lamba} from cross-validation
\item \code{bestTune} values of selected hyperparameter during cross-validation (i.e, \code{labmda} value)
}
}
