% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_getPatterns.R
\name{getPatIDs}
\alias{getPatIDs}
\title{Get patient IDs.}
\usage{
getPatIDs(ids)
}
\arguments{
\item{ids}{an array of ids in the form "patientID.eventID"}
}
\value{
an array of patientIDs
}
\description{
Get patient IDs.
}
\details{
Given a string in the form "patientID.eventID", return just the patientID.
}
\examples{
data("features_ratechange_sup0.4g60l2z2")
feats$patID <- getPatIDs(rownames(feats))
}
