% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_control.R
\name{getPerformanceNames}
\alias{getPerformanceNames}
\title{Get colnames that will be kept for averaging across cross validation.}
\usage{
getPerformanceNames()
}
\value{
char
}
\description{
Get colnames that will be kept for averaging across cross validation.
}
\details{
A results dataframe is returned from caret. These are the column
names that are numeric and can be averaged across multiple results.
}
\seealso{
\code{\link{aucs}}
}
