% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_performance.R
\name{getROC_PR}
\alias{getROC_PR}
\title{depends on ROCR}
\usage{
getROC_PR(prob, round = TRUE)
}
\arguments{
\item{prob}{An object of class prediction, predicted values and ground truth, input to \code{\link[ROCR]{performance}}}

\item{round}{logical, whether to round roc and pr axis values to 3 decimals.}
}
\value{
list of dataframes \code{roc} and \code{pr} x and y data points, and the area under the curves.
}
\description{
Get the ROC and PR curve from model probabilies.
}
