% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_getPatterns.R
\name{getSeqPatterns}
\alias{getSeqPatterns}
\title{Get sequential patterns}
\usage{
getSeqPatterns(event, transFilename, createT, support, maxgap, maxlen, maxsize)
}
\arguments{
\item{event}{dataframe of events,
where rows are single events, and
least the following columns:
\itemize{
\item \code{iois} patient id
\item \code{eventID} event id, see \code{\link{resetEventIDsToDays}}
\item \code{eventName} unique event name, see \code{\link{createEventNames}}
}}

\item{transFilename}{asdf}

\item{createT}{logical, create or load previously created transactions}

\item{support}{see slot in \code{\link[arulesSequences]{SPparameter-class}}}

\item{maxgap}{see slot in \code{\link[arulesSequences]{SPparameter-class}}}

\item{maxlen}{see slot in \code{\link[arulesSequences]{SPparameter-class}}}

\item{maxsize}{see slot in \code{\link[arulesSequences]{SPparameter-class}}}
}
\value{
a named list containing: \itemize{
\item freqseq, the output from \code{\link{arulesSequences::cspade}}
\item data, the input \code{data}
}
}
\description{
Get sequential patterns
}
\examples{
data("fake_data")
outputDir <- '~/test'

# save tumor location and laterility strings before event cleaning
fake_tumorInfo <- fake_data$events 
fake_demo <- fake_data$demo

# clean data
fake_data$events <- cleanData(fake_data$events, tType = 'rate')
cat('...',nrow(fake_data$events), " events left for SPM after cleaning", '\\n')

# collect patient info for each event
fake_data <- merge(fake_data$events, fake_data$person, by='iois', all.x=T) 

# prep for each event, since age does change
# get survival labels, these also change
fake_data <- prepDemographics(fake_data, fake_demo) 
fake_data <- prepSurvivalLabels(fake_data) 

# get first tumor location
fake_data <- getTumorLocation(fake_data, fake_tumorInfo) 

# spm
pSPM <- getSeqPatterns(event = fake_data,
                       transFilename = file.path(outputDir, 'example_transactions.txt'),
                       createT = T,
                       support = 0.4,
                       maxgap = 60,
                       maxlen = 2,
                       maxsize = 2)
}
\seealso{
\code{\link{createTransactions}}
}
