% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_partition.R
\name{getTrainingFolds}
\alias{getTrainingFolds}
\title{Create patient-independent folds in cross-validation in training partition.}
\usage{
getTrainingFolds(trainEvents, folds, seed, verbose)
}
\arguments{
\item{trainEvents}{a dataframe holding the event data ready for logit modeling, 
where each row is an event/clinical visit and the columns
contain features of the event and the labels. It must contain an \code{id} column 
in the form "patientID.eventID".}

\item{folds}{number of folds}

\item{seed}{int, seed for split}

\item{verbose}{logical, True for print, False for silence}
}
\value{
\code{folds} a list of arrays indicating row 
position integers corresponding to fold split
}
\description{
Create patient-independent folds in cross-validation in training partition.
}
\details{
Patients in validation folds will not have events in training folds 
and vice versa.
}
\examples{

# use training partition to create folds for CV
data("features_ratechange_sup0.4g60l2z2") # features and labels for each clinical visit
t <- 'rate'
maxgap <- 60
maxlen <- 2
names <- colnames(feats)
feats <- data.frame(id=row.names(feats),feats)
colnames(feats) <- c('id',names)
feats <- prepLaterality(feats)
feats <- prepLocation(feats)
feats <- removeVisits(feats,
                     maxgap=maxgap,
                     maxlength=maxlen,
                     tType=t,
                     save=F,
                     outDir=NA)
labels <- getClassLabels()
needToRemove <- c('id','iois','eventID', # remove ids
                  labels, # remove labels
                  'IDH1') # not interested

# data partitions
train.ids <- sample(feats$id, size=floor(0.80*nrow(feats)), replace = F) # random
feats <- feats[feats$id \%in\% train.ids,] #training data
ind <- getTrainingFolds(trainEvents=feats,
                        folds=5,
                        seed=1,
                        verbose=T)
                        
}
\seealso{
\code{\link[caret]{groupKFold}},
\code{\link{getPatIDs}},
\code{\link{getClassLabels}},
\code{\link{getData}}
}
