% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_control.R
\name{getTrainResults}
\alias{getTrainResults}
\title{get training results of fitted model retrained on entire training partition}
\usage{
getTrainResults(model)
}
\arguments{
\item{model}{glmnet fitted}
}
\description{
asdf
}
