% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_getPatterns.R
\name{getTumorNames}
\alias{getTumorNames}
\alias{getDemoNames}
\alias{getBiomarkerNames}
\alias{getClassLabels}
\title{Get column names of labels and non-spm features to potentially include in logit fitting data.}
\usage{
getTumorNames()

getDemoNames()

getBiomarkerNames()

getClassLabels()
}
\value{
char, names of features and labels to look for in fitting data
}
\description{
Get column names of labels and non-spm features to potentially include in logit fitting data.
}
\examples{
nonPatternColNames <- c('iois','eventID', # identification
                        getDemoNames(),
                        getBiomarkerNames(), 
                        getTumorNames(),
                        getClassLabels())
}
