% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{getVolTypeName}
\alias{getVolTypeName}
\title{Get the full name of the tumor volume type.}
\usage{
getVolTypeName(tType)
}
\arguments{
\item{tType}{char, shortcut for type of tumor volume information}
}
\value{
full name of tumor volume type
}
\description{
Map shortcut name to fullname for type of tumor volume information.
Acceptable types: \itemize{
        \item 'c' for continuous, left as is, e.g., volume was measured in cubic mm.
        \item 'p' for percent change
        \item 'r' for volumetric response criteria
        \item 'rate' for rate change
        \item 'rc' for both 'c' and 'rate'
}
See \code{prepTumorVol} for full details.
}
\examples{
getVolTypeName('p')
getVolTypeName('r')
getVolTypeName('rate')

}
