% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_performance.R
\name{plotMultiROC}
\alias{plotMultiROC}
\alias{plotMultiPR}
\title{Multiple ROC plots in one and also faceted}
\usage{
plotMultiROC(fn, rocsList, aucsList, listOrder, legendTitle, printIt, outDir,
  perfTheme, scaleX, scaleY, perfGuide)

plotMultiPR(fn, prsList, aucsList, listOrder, legendTitle, printIt, outDir,
  perfTheme, scaleX, scaleY, perfGuide)
}
\arguments{
\item{fn}{plot filename}

\item{rocsList}{named list of ROCs, each containing dataframe with columns:
\itemize{
\item \code{x} false positive rate
\item \code{y} true positive rate
\item \code{L1} names of different ROCs to differeniate x and y (important if there are multiple ROCs, i.e., training vs testing)
}}

\item{aucsList}{named list of area under the ROC values}

\item{listOrder}{order of plotting for legend labeling of multipe ROC or PR curves}

\item{outDir}{directory to put \code{fn}, the plot}

\item{perfTheme}{ggplot2 theme() object to include plot formats}

\item{scaleX}{scale_x_continuous, format of x scale}

\item{scaleY}{scale_y_continuous, format of y scale}

\item{prsList}{named list of  precision-recall curves, each containing dataframe with columns:
\itemize{
\item \code{x} recall
\item \code{y} precision
\item \code{L1} names of different ROCs to differeniate x and y (important if there are multiple ROCs, i.e., training vs testing)
}}
}
\description{
Plots multiple ROC or PR curves in one plot (i.e., different models), 
and faceted by another variable (i.e., different classification task)
}
