% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_performance.R
\name{plotROC}
\alias{plotROC}
\alias{plotPR}
\title{Plot multiple performance curves in one graph}
\usage{
plotROC(ptitle, fn, rocs, aucs, names, outDir, perfTheme, scaleX, scaleY,
  scaleColor)

plotPR(ptitle, fn, prs, aucs, names, outDir, perfTheme, scaleX, scaleY,
  scaleColor)
}
\arguments{
\item{ptitle}{plot title}

\item{fn}{plot filename}

\item{rocs}{list of ROCs, each containing dataframe with columns:
\itemize{
\item \code{x} false positive rate
\item \code{y} true positive rate
\item \code{L1} names of different ROCs to differeniate x and y (important if there are multiple ROCs)
}}

\item{aucs}{area under the ROC values matching the order of \code{names}}

\item{names}{names of the different ROCs in \code{L1}, to be included in legend name}

\item{outDir}{directory to put \code{fn}, the plot}

\item{perfTheme}{ggplot2 theme() object to include plot formats}

\item{scaleX}{scale_x_continuous, format of x scale}

\item{scaleY}{scale_y_continuous, format of y scale}

\item{scaleColor}{scale_color_manual, specify colors}

\item{prs}{list of precision-recall curves, each containing dataframe with columns:
\itemize{
\item \code{x} recall
\item \code{y} precision
\item \code{L1} names of different curvers to differeniate x and y (important if there are multiple)
}}
}
\description{
Plot multiple performance curves in one graph
}
