% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{prepDemographics}
\alias{prepDemographics}
\title{Add and prep demographic information (static 'events') to events.}
\usage{
prepDemographics(data, demo, ageCategories = c(seq(20, 70, 10), 90))
}
\arguments{
\item{data}{dataframe, 
rows are clinical visits to be classified, 
columns are features including clinical visits ids 
and at least the following columns:
\itemize{
\item \code{iois} char, patient id
\item \code{startdate} = event start date in the format 'YYYY-MM-DD', char
}}

\item{demo}{dataframe, patient demograpics and biomarker information, containing columns:
\itemize{
\item \code{iois} char, patient id
\item \code{ethnicity} char, patient ethnicity, uses value 'WHITE' to differentiate into three categories: WHITE, non-WHITE, and unknown
\item \code{DOB} char, date of birth in the format 'YYYY-MM-DD'
\item \code{MSP} char, MGMT promoter methytation status, 'U' = unmethylated, 'M' = methylated
}
returns
- data: dataframe, with formated demographics}
}
\description{
Ethnicity is either 'white' or 'non-white'.
Age is broken down to decades, with 70-90 decades merged.
}
\examples{
data("fake_data")
t <- 'rate'
fake_tumorInfo <- fake_data$events # save tumor location and laterility strings before event cleaning
fake_demo <- fake_data$demo

fake_data$events <- cleanData(fake_data$events, tType = t)

# collect patient info for each event
# note that only eventName, iois, and eventID columns are used for SPM
fake_data <- merge(fake_data$events, fake_data$person, by='iois', all.x=T)
fake_data <- prepDemographics(fake_data, fake_demo) # prep for each event, since age does change

}
