% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{prepLaterality}
\alias{prepLaterality}
\title{Prepare tumor laterality feature for logit modeling.}
\usage{
prepLaterality(data)
}
\arguments{
\item{data}{dataframe, 
rows are clinical visits to be classified, 
columns are features including clinical visits ids,
and at least the following column:
\itemize{
\item \code{tumorLaterality} char, tumor laterality
}}
}
\value{
\code{data} dataframe with replaced laterality columns
}
\description{
Prepare tumor laterality feature for logit modeling.
}
\details{
Converts tumor laterality feature into two boolean features: "left" and "right".
Searches for "left", "right", "both" in \code{tumorLaterality} 
and sets "both" as "left" and "right."
}
\examples{

data("features_ratechange_sup0.4g60l2z2")  # features and labels for each clinical visit

# format
names <- colnames(feats)
feats <- data.frame(id=row.names(feats),feats)
colnames(feats) <- c('id',names)
cat('...overall samples: ', nrow(feats), '\\n')

feats <- prepLaterality(feats) 

}
