% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{prepLocation}
\alias{prepLocation}
\title{Prepare tumor location feature for logit modeling.}
\usage{
prepLocation(data)
}
\arguments{
\item{data}{dataframe, 
rows are clinical visits to be classified, 
columns are features including clinical visits ids,
and at least the columns returned from \code{\link{getLocation}}}
}
\value{
\code{data} dataframe with replaced location columns
}
\description{
Locations with few observations are merged into 'other'. 
Categories include: frontal, temporal, parietal, and other.
}
\examples{

data("features_ratechange_sup0.4g60l2z2")  # features and labels for each clinical visit

# format
names <- colnames(feats)
feats <- data.frame(id=row.names(feats),feats)
colnames(feats) <- c('id',names)
cat('...overall samples: ', nrow(feats), '\\n')

feats <- prepLaterality(feats) 

}
