% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{prepSurvivalLabels}
\alias{prepSurvivalLabels}
\title{Prep survival information (labels for all classifictation tasks) to events.}
\usage{
prepSurvivalLabels(data)
}
\arguments{
\item{data}{dataframe, 
rows are clinical visits to be classified, 
columns are features including clinical visits ids 
and at least the following columns:
\itemize{
\item \code{eventID} num, event id
\item \code{survival} num, days since baseline
\item \code{status} char, patient status at the end of data collection, 'DECEASED' or 'ALIVE'
}}
}
\value{
data dataframe, 
rows are clinical visits to be classified, 
columns are features including clinical visits ids 
and the following added columns:
\itemize{
\item \code{daysToDeath} num, event id
\item \code{survivalIn60} char, 'dead' or 'alive' in 60 days or ~2 months
\item \code{survivalIn180} char, 'dead' or 'alive' in 180 days or ~6 months
\item \code{survivalIn270} char, 'dead' or 'alive' in 270 days or ~9 months
\item \code{survivalIn360} char, 'dead' or 'alive' in 360 days or ~12 months
}
}
\description{
Prep survival information (labels for all classifictation tasks) to events.
}
\examples{
data("fake_data")
t <- 'rate'
fake_tumorInfo <- fake_data$events # save tumor location and laterility strings before event cleaning
fake_demo <- fake_data$demo

fake_data$events <- cleanData(fake_data$events, tType = t)

# collect patient info for each event
# note that only eventName, iois, and eventID columns are used for SPM
fake_data <- merge(fake_data$events, fake_data$person, by='iois', all.x=T)
fake_data <- prepDemographics(fake_data, fake_demo) # prep for each event, since age does change
fake_data <- prepSurvivalLabels(fake_data) # get survival labels, these also change

}
