% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{prepTumorVol}
\alias{prepTumorVol}
\title{Get all tumor volume types.}
\usage{
prepTumorVol(volumes)
}
\arguments{
\item{volumes}{a dataframe containing  tumor measurement events acrosss patients, see Details}
}
\value{
\code{volumes} dataframe, with additional columns for each type of tumor volume information, see Details
}
\description{
Get all tumor volume types.
}
\details{
\code{volumes} contains tumor measurement events as rows and at least the following columns: \itemize{
\item iois = patient id, char
\item agent_detail = event attribute, i.e., measured tumor volume in mm3 char
\item startdate = event start date in the format 'YYYY-MM-DD', char
\item eventName = event name, char
}

Tumor volume types are: \itemize{
\item volume = volume in original measured units (e.g., mm^3)
\item baseline size = earliest tumor volume observed in a patient
\item percent change = \eqn{(Volume_{current} - Volume_{previous})/Volume_{previous} }
\item rate change = \eqn{(Volume_{current} - Volume_{previous})/(DaysSinceBaseline_{current} - DaysSinceBaseline_{previous}) }
}
}
\examples{
data("fake_data")
fake_demo <- fake_data$demo
fake_tumorInfo <- fake_data$events

# by default eventName is the volumetric response critera
# will get all types eventually
fake_data$events <- cleanData(fake_data$events) 

# keep tumor volume only
fake_data$events <- fake_data$events[fake_data$events$agent=='measurement',] 
fake_data$events <- prepTumorVol(fake_data$events) # get all tumor vol types

}
