% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{removeData}
\alias{removeData}
\title{Formats event strings and removes incorrect or missing information.}
\usage{
removeData(event)
}
\arguments{
\item{event}{a dataframe containing events, see Details}
}
\value{
the \code{event} dataframe with removed events
}
\description{
Formats event strings and removes incorrect or missing information.
}
\details{
\code{event} contains events as rows and at least the following columns:
\itemize{
\item agent = event type, char
\item agent_detail = event attribute, char
}
}
\examples{
data("fake_data")
events <- fake_data$events
events$startdate <- as.Date(as.character(events$startdate), format='\%Y-\%m-\%d')
events$enddate <- as.Date(as.character(events$enddate), format='\%Y-\%m-\%d')

cleaned <- removeData(events)

}
