% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{removePreRadiation}
\alias{removePreRadiation}
\title{Remove events which occurred prior to a patient's first radiation}
\usage{
removePreRadiation(event)
}
\arguments{
\item{event}{a dataframe containing all events across all patients (assumes more than 1 patient), see Details}
}
\value{
the \code{event} dataframe with removed 'Radiation' events
}
\description{
Remove events which occurred prior to a patient's first radiation
}
\details{
\code{event} contains events as rows and at least the following columns:
\itemize{
\item iois = patient id, char
\item agent = event type, char
\item eventID = event id, num
}

Assumes radiation is labelled as 'Radiation' in \code{agent}.
}
\examples{
data("fake_data")
events <- fake_data$events
events$startdate <- as.Date(as.character(events$startdate), format='\%Y-\%m-\%d')
events$enddate <- as.Date(as.character(events$enddate), format='\%Y-\%m-\%d')

cleaned <- removeData(events)
cleaned <- resetEventIDsToDays(cleaned)
cleaned <- removePreRadiation(cleaned)

}
