% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{resetEventIDsToDays}
\alias{resetEventIDsToDays}
\title{Convert event dates into relative days.}
\usage{
resetEventIDsToDays(event)
}
\arguments{
\item{event}{a dataframe containing all events across all patients, see Details}
}
\value{
the \code{event} dataframe with additional columns: and added columns: \itemize{
\item \code{eventID}
\item \code{endEventID}
}
}
\description{
Set \code{eventID} relative to each person. 
\code{eventID} will be in days after the first observation date.
}
\details{
\code{event} contains events as rows and at least the following columns:
\itemize{
\item iois = patient id, char
\item startdate = event start date in the format 'YYYY-MM-DD', char
\item enddate = event end date in the format 'YYYY-MM-DD', char
}
Relative days refer to the number of days since baseline.
Baseline date is the earliest event observed in a patient.

\code{eventID}: number of days since first event (grouped by patient id) to the start date of event
\code{endEventID}: number of days since first event (grouped by patient id) to the end date of event
}
\examples{
data("fake_data")
events <- fake_data$events
events$startdate <- as.Date(as.character(events$startdate), format='\%Y-\%m-\%d')
events$enddate <- as.Date(as.character(events$enddate), format='\%Y-\%m-\%d')

cleaned <- removeData(events)
cleaned <- resetEventIDsToDays(cleaned)

}
