% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_control.R
\name{retrainLogit}
\alias{retrainLogit}
\title{Retrain logit after hyperparameter selection.}
\usage{
retrainLogit(data, formula, labelName, lambda = NULL, lasso = TRUE,
  predictors = NULL, needToRemove = NULL, createModelMatrix = FALSE,
  metric = c("prAUC", "rocAUC"), seed, verbose = TRUE)
}
\arguments{
\item{data}{dataframe, rows are samples, cols are features plus some metadata
not meant for modeling and will be removed}

\item{formula}{char or formula object}

\item{labelName}{char, column name of binary label}

\item{lambda}{numberic, lambda value}

\item{lasso}{logical, whether to use lasso regularization}

\item{predictors}{char, names of columns in \code{data} that should be in logit fit data}

\item{needToRemove}{char, names of columns in \code{data} that should not be in logit fit data}

\item{createModelMatrix}{logical, call \code{\link[stats]{model.matrix}}}

\item{metric}{char, see \code{aucs}}

\item{seed}{int, seed for split}

\item{verbose}{logical, True for print, False for silence}
}
\description{
Used for fitting the full training partition.
}
