% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_control.R
\name{runCV}
\alias{runCV}
\title{Run logit modeling}
\usage{
runCV(data, cvReps, formula, labelName, lasso = FALSE, llength = NULL,
  lmax = NULL, predictors = NULL, needToRemove = NULL,
  createModelMatrix = FALSE, metric = c("prAUC", "rocAUC"), seed, folds)
}
\arguments{
\item{data}{dataframe, rows are samples, cols are features plus some metadata
not meant for modeling and will be removed}

\item{formula}{char or formula object}

\item{labelName}{char, column name of binary label}

\item{lasso}{logical, whether to use lasso regularization}

\item{llength}{num, number of lambdas to consider up to \code{lmax}}

\item{lmax}{num, maximum lambda to consider, cannot be NULL if lambda is NULL}

\item{predictors}{char, names of columns in \code{data} that should be in logit fit data}

\item{needToRemove}{char, names of columns in \code{data} that should not be in logit fit data}

\item{createModelMatrix}{logical, call \code{\link[stats]{model.matrix}}}

\item{metric}{char, see \code{aucs}}

\item{seed}{int, seed for split}

\item{folds}{number of folds}
}
\value{
list of logit fits, averaged cross-validation results (if \code{cvReps} > 1), and test data partition 
(one that is formated like training data and is ready to be called by subsequent trained model)
}
\description{
Patients are first split into training and testing partitions. 
Next, samples with NA features will be removed. 
Then, training partition is split for cross-validation so that
no patient has events in both validation and training. 
Both partitions are prepped for logit fitting and cross-validation is run.
}
\seealso{
\code{\link{prepLogitData}}, 
\code{\link{getCV}}, 
\code{\link{getPerformanceNames}}
}
