% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm_getPatterns.R
\name{runSPM}
\alias{runSPM}
\title{Run sequential pattern mining}
\usage{
runSPM(event, suppList, maxgap, maxlen, maxsize, tType, outputDir, createT)
}
\arguments{
\item{event}{dataframe of events,
where rows are single events, and
least the following columns:
\itemize{
\item \code{iois} patient id
\item \code{eventID} event id, see \code{\link{resetEventIDsToDays}}
\item \code{eventName} unique event name, see \code{\link{createEventNames}}
}}

\item{suppList}{one or more supports to be used as a parameter, see slot in
code{\link[arulesSequences]{SPparameter-class}}}

\item{maxgap}{see slot in \code{\link[arulesSequences]{SPparameter-class}}}

\item{maxlen}{see slot in \code{\link[arulesSequences]{SPparameter-class}}}

\item{maxsize}{see slot in \code{\link[arulesSequences]{SPparameter-class}}}

\item{tType}{char, shortcut for type of tumor volume information}

\item{outputDir}{full path to output folder, e.g., where to save spm results}

\item{createT}{logical, create or load previously created transactions}
}
\description{
Run sequential pattern mining
}
\details{
This will automatically create three types of files: \itemize{
\item transactions - returned from \code{\link{createTransactions}}, saved in .txt
\item patterns - returned from \code{\link{findPatternDays}}, saved in .rds 
\item features - returned from \code{\link{getFeatureVectors}}, saved in .rds 
}
}
\examples{
data("fake_data")
outputDir <- '~/test'

# save tumor location and laterility strings before event cleaning
fake_tumorInfo <- fake_data$events 
fake_demo <- fake_data$demo

# clean data
fake_data$events <- cleanData(fake_data$events, tType = 'rate')
cat('...',nrow(fake_data$events), " events left for SPM after cleaning", '\\n')

# collect patient info for each event
fake_data <- merge(fake_data$events, fake_data$person, by='iois', all.x=T) 

# prep for each event, since age does change
# get survival labels, these also change
fake_data <- prepDemographics(fake_data, fake_demo) 
fake_data <- prepSurvivalLabels(fake_data) 

# get first tumor location
fake_data <- getTumorLocation(fake_data, fake_tumorInfo) 

runSPM(event=fake_data,
       suppList=suppList, 
       maxgap=60, 
       maxlen=2, 
       maxsize=2, 
       tType='rate', 
       outputDir='~/test')

}
\seealso{
\code{\link{getSeqPatterns}}, \code{\link{getVolTypeName}}
}
