% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_performance.R
\name{selectROCThreshold}
\alias{selectROCThreshold}
\title{Select the roc threshold with the max F1 score and calculate classification metrics}
\usage{
selectROCThreshold(trainPred)
}
\arguments{
\item{trainPred}{An object of class prediction, predicted values and ground truth, input to \code{\link[ROCR]{performance}}}
}
\value{
dataframe of various metrics when using threshold to do classification, 
including sensitivity, specificity, TP, FP, TN, FN, precision, recall, f1score
}
\description{
Select the roc threshold with the max F1 score and calculate classification metrics
}
