% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_discretize.R
\name{setPercentChange}
\alias{setPercentChange}
\alias{setRatechange}
\alias{setResponse}
\title{Calculate different tumor volume information types.}
\usage{
setPercentChange(volumes)

setRatechange(volumes)

setResponse(volumes)
}
\arguments{
\item{volumes}{dataframe, a single patient's tumor volume events}
}
\value{
\code{volumes} with \code{percentChange} as an additional column
}
\description{
\code{setPercentChange}: get volume percent changes from a sequence of volume measurements.

\code{setRatechange}: get volume rate changes from a sequence of volume measurements.

\code{setResponse}: get volumetric response criteria by mapping continuous volumes to response criteria categories
}
\details{
Percent changes, rate changes, and volumetric response criteria (see \code{prepTumorVol}) 
are based on volumes in \code{volumes}, where rows are ordered (e.g,. by time).

\code{volumes} contains tumor measurement events as rows and at least the following column: \itemize{
\item agent_detail = tumor volume in original units (e.g, cubic mm)
\item eventID = baseline day of occurence, accepts only one volume per day
}
}
\examples{
data("fake_data")
vols <- fake_data$events[fake_data$events$agent=='measurement', ]

# format data
vols$startdate <- as.Date(as.character(vols$startdate), format='\%Y-\%m-\%d')
vols$enddate <- as.Date(as.character(vols$enddate), format='\%Y-\%m-\%d')
vols <- resetEventIDsToDays(vols)
vols <- removeData(vols) # get rid of excess details and empty or no data

# get vols for one patient
vols_set <- vols[vols$iois=='1',] 
p <- setPercentChange(vols_set)

data("fake_data")
vols <- fake_data$events[fake_data$events$agent=='measurement', ]

# format data
vols$startdate <- as.Date(as.character(vols$startdate), format='\%Y-\%m-\%d')
vols$enddate <- as.Date(as.character(vols$enddate), format='\%Y-\%m-\%d')
vols <- resetEventIDsToDays(vols)
vols <- removeData(vols) # get rid of excess details and empty or no data

vols_set <- vols[vols$iois=='55',]
r <- setRatechange(vols_set)

data("fake_data")
vols <- fake_data$events[fake_data$events$agent=='measurement', ]

# format data
vols$startdate <- as.Date(as.character(vols$startdate), format='\%Y-\%m-\%d')
vols$enddate <- as.Date(as.character(vols$enddate), format='\%Y-\%m-\%d')
vols <- resetEventIDsToDays(vols)
vols <- removeData(vols) # get rid of excess details and empty or no data

vols_set <- vols[vols$iois=='55',]
r <- setResponse(vols_set)

}
